+++ src/dataprotection/azext_dataprotection/aaz/latest/dataprotection/backup_instance/deleted_backup_instance/__init__.py
@@ -8,17 +8,7 @@
 # pylint: skip-file
 # flake8: noqa
 
-from azure.cli.core.aaz import *
-
-
-@register_command_group(
-    "dataprotection backup-instance",
-    is_experimental=True,
-)
-class __CMDGroup(AAZCommandGroup):
-    """Manage backup instance with dataprotection.
-    """
-    pass
-
-
-__all__ = ["__CMDGroup"]
+from .__cmd_group import *
+from ._list import *
+from ._show import *
+from ._undelete import *